Custom confirm dialog window added to client

+++ clientApp/src/app/members/member-messages/member-messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { Message } from 'src/app/_models/message';
+import { ConfirmWindowService } from 'src/app/_services/confirm-window.service';
 import { MessageService } from 'src/app/_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MemberMessagesComponent implements OnInit {
   messageContent: string;
   loading = false;
 
-  constructor(public messageService: MessageService) { }
+  constructor(public messageService: MessageService, private confirmWindowService: ConfirmWindowService) { }
 
   ngOnInit(): void {    
   }
@@ -30,8 +31,12 @@ export class MemberMessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number, deleteBoth:string){
-    this.messageService.deleteMessage(id, deleteBoth).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id),1);
-    })
+    this.confirmWindowService.confirm('Delete Message?','This will not undone after').subscribe(result => {
+      if (result){
+        this.messageService.deleteMessage(id, deleteBoth).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id),1);
+        })
+      }
+    })    
   }
 }
